Bug 1465589 - convert math and compare service to helper functions
