+++ ui/entry-perf.js
@@ -27,11 +27,11 @@ import './js/filters';
 import './js/models/perf/issue_tracker';
 import './js/models/perf/performance_framework';
 import './js/models/perf/alerts';
-import './js/services/perf/math';
-import './js/services/perf/compare';
 import './js/controllers/perf/compare';
 import './js/controllers/perf/graphs';
 import './js/controllers/perf/alerts';
 import './js/components/perf/compare';
 import './js/components/loading';
 import './js/perfapp';
+import './perfherder/constants';
+import './perfherder/helpers';
