--- ui/js/services/perf/math.js
@@ -1,78 +1,0 @@
-// Remove the eslint-disable when rewriting this file during the React conversion.
-/* eslint-disable func-names, object-shorthand, prefer-arrow-callback */
-import treeherder from '../../treeherder';
-
-treeherder.factory('math', [
-    function () {
-        function percentOf(a, b) {
-            return b ? 100 * a / b : 0;
-        }
-
-        function average(values) {
-            if (values.length < 1) {
-                return 0;
-            }
-
-            return values.reduce((a, b) => a + b, 0) / values.length;
-        }
-
-        function stddev(values, avg) {
-            if (values.length < 2) {
-                return undefined;
-            }
-
-            if (!avg) avg = average(values);
-
-            return Math.sqrt(
-                values.map(v => (v - avg) ** 2)
-                    .reduce(function (a, b) { return a + b; }) / (values.length - 1));
-        }
-
-        // If a set has only one value, assume average-ish-plus sddev, which
-        // will manifest as smaller t-value the less items there are at the group
-        // (so quite small for 1 value). This default value is a parameter.
-        // C/T mean control/test group (in our case original/new data).
-        function t_test(valuesC, valuesT, stddev_default_factor) {
-            const lenC = valuesC.length;
-            const lenT = valuesT.length;
-
-            // We must have at least one value at each set
-            if (lenC < 1 || lenT < 1) {
-                return 0;
-            }
-
-            const avgC = average(valuesC);
-            const avgT = average(valuesT);
-
-            // Use actual stddev if possible, or stddev_default_factor if one sample
-            let stddevC = (lenC > 1 ? stddev(valuesC, avgC) : stddev_default_factor * avgC);
-            let stddevT = (lenT > 1 ? stddev(valuesT, avgT) : stddev_default_factor * avgT);
-
-            // If one of the sets has only a single sample, assume its stddev is
-            // the same as that of the other set (in percentage). If both sets
-            // have only one sample, both will use stddev_default_factor.
-            if (lenC === 1) {
-                stddevC = valuesC[0] * stddevT / avgT;
-            } else if (lenT === 1) {
-                stddevT = valuesT[0] * stddevC / avgC;
-            }
-
-            const delta = avgT - avgC;
-            const stdDiffErr = (
-                Math.sqrt(
-                    stddevC * stddevC / lenC // control-variance / control-size
-                        +
-                        stddevT * stddevT / lenT, // ...
-                )
-            );
-
-            return delta / stdDiffErr;
-        }
-
-        return {
-            percentOf: percentOf,
-            average: average,
-            stddev: stddev,
-            t_test: t_test,
-        };
-    }]);
